requirements.txt
Django>=5.0.0,<7.0.0
djangorestframework>=3.14.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
